nvim-tree: group empty folders, slight upgrade
decided to enable the nvim-tree feature to group folders containing only
a subfolder together, to make handier to work in java projects. That
made me hit https://github.com/nvim-tree/nvim-tree.lua/issues/1716 in my
1-year-old version of nvim-tree. Tried to upgrade to the latest
nvim-tree and put the config changes that enable that, since nvim-tree
changed its setup, but had to revert due to
http://github.com/stevearc/stickybuf.nvim/issues/24

So in the end keep the code to migrate to the latest nvim-tree, but
commented, and do only a slight nvim-tree upgrade to fix the folder
grouping issue.